feat: Experiment with native LSP completion
This extends previous experiments with triggering the LSP completions
manually. It now adopts a native approach to trigger the completion and
leverage the classic popup menu together with fuzzy matching for
completion selection. All completion is now triggered manually through
'Ctrl-N', which either triggers the LSP completion whenever an LSP is
attached to the buffer or the class omni completion if no LSP is
present.

This does lose some functionality, such as the ability to show signature
directly from within the completions, but it seems more light weight and
responsive. Also, in practise it seems that we hardly use all
functionality provided by common completion implementation and the
current approach might just be sufficient.

Since this is somewhat experimental and requires version 11 that is not
yet out, it will be kept on this branch for a while.